grouped logic in one file and cloud support instead of localhost
